server/requirements.txt
flask
pymongo
flask-pymongo
python-dotenv
authlib
requests
pytest